# Improving software performance of AES-GCM for unaligned AAD and plaintext sizes