Hello everyone,

This commit is a contribution to OpenSSL.

It improves the performance of the AES-GCM for specific AAD and plaintext
(PT) sizes, while almost not degrading the perofrmance for the other
AAD/PT sizes.

It improves the performance for short inputs (upto 512 bytes) |AAD| mod 16
or |PT| mod 16 >= 8. Depending on the inputs sizes, one can observe upto
1.14x speedup. To facilitate this observation, the code of speed.c
is modified, to include various PT lengths with a fixed 13-byte AAD length
(as defined in TLS 1.2). In addition, we changed the code to measure the entire
AES-GCM algorithm and not only the incremental update (encryption) parts.
The modified speed.c is included in the patch, for testing purposes.

A full report is posted in [1].

Examples:
Current OpenSSL:
Doing aes-256-gcm for 1s on 2 size blocks: 5004044 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 4 size blocks: 8357968 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 6 size blocks: 8266176 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 8 size blocks: 8172821 aes-256-gcm's in 1.00s

Optimzed OpenSSL:
Doing aes-256-gcm for 1s on 2 size blocks: 5033976 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 4 size blocks: 8624260 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 6 size blocks: 8487607 aes-256-gcm's in 1.00s
Doing aes-256-gcm for 1s on 8 size blocks: 8618943 aes-256-gcm's in 1.00s

Reference:
[1] N. Drucker, S. Gueron, Improving software performance of AES-GCM for
    unaligned AAD and plaintext sizes, eprint (2018).

Developers and authors:
***************************************************************************
Nir Drucker (1,2), Shay Gueron (1,2)
(1) Amazon Web Services Inc.
(2) University of Haifa, Israel
***************************************************************************
Copyright (c) 2018, Amazon Web Services Inc.

<!--
Thank you for your pull request. Please review these requirements:

Contributors guide: https://github.com/openssl/openssl/blob/master/CONTRIBUTING

Other than that, provide a description above this comment if there isn't one already

If this fixes a github issue, make sure to have a line saying 'Fixes #XXXX' (without quotes) in the commit message.
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->
- [ ] documentation is added or updated
- [ ] tests are added or updated