@richsalz 2018-05-16
We need an update to the Amazon CLA that includes both of you, and we need CLA's from each of you with the proper email addresses added.  Thanks!